requirements.txt
# Core dependencies
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
# sqlite3

# Analytics Dashboard
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0

# Inference (optional)
openai>=1.10.0
aiohttp>=3.9.0
pyyaml>=6.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0

# Development
ruff>=0.1.0
mypy>=1.8.0

# Monitoring
prometheus-client>=0.20.0
